requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.6
Django==4.1.5
django-allauth==0.52.0
django-channels==0.7.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
oauthlib==3.2.2
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.0.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
zope.interface==5.5.2
